delegator: use either direct or indirect proxying
