+++ delegator/proxy_test.go
@@ -17,6 +17,7 @@ func Test_newProxy(t *testing.T) {
 	type args struct {
 		feature string
 		target  *url.URL
+		direct  bool
 	}
 	type request struct {
 		method  string
@@ -30,22 +31,26 @@ func Test_newProxy(t *testing.T) {
 		want *url.URL
 	}{
 		{"api redirect with GET",
-			args{"api", &url.URL{Scheme: "http", Host: "127.0.0.1"}},
+			args{"api", &url.URL{Scheme: "http", Host: "127.0.0.1"}, false},
 			request{"GET", "http://127.0.0.1/networks/blah/api/v0/blah", nil},
 			&url.URL{Scheme: "http", Host: "127.0.0.1", Path: "/api/v0/blah"}},
 		{"api redirect with POST",
-			args{"api", &url.URL{Scheme: "http", Host: "127.0.0.1"}},
+			args{"api", &url.URL{Scheme: "http", Host: "127.0.0.1"}, false},
 			request{"POST", "http://127.0.0.1/networks/blah/api/v0/blah", bytes.NewReader(testBytes)},
 			&url.URL{Scheme: "http", Host: "127.0.0.1", Path: "/api/v0/blah"}},
-		{"default redirect",
-			args{"swarm", &url.URL{Scheme: "http", Host: "127.0.0.1"}},
+		{"indirect redirect",
+			args{"swarm", &url.URL{Scheme: "http", Host: "127.0.0.1"}, false},
 			request{"GET", "http://127.0.0.1/networks/blah/swarm/blah", nil},
+			&url.URL{Scheme: "http", Host: "127.0.0.1", Path: "/blah"}},
+		{"direct redirect",
+			args{"swarm", &url.URL{Scheme: "http", Host: "127.0.0.1"}, true},
+			request{"GET", "https://node.gateway.temporal.cloud/blah", nil},
 			&url.URL{Scheme: "http", Host: "127.0.0.1", Path: "/blah"}},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			var l = zaptest.NewLogger(t).Sugar()
-			var got = newProxy(tt.args.feature, tt.args.target, l)
+			var got = newProxy(tt.args.feature, tt.args.target, l, tt.args.direct)
 
 			// test proxy Director
 			var r = httptest.NewRequest(tt.req.method, tt.req.address, tt.req.body)
@@ -62,6 +67,27 @@ func Test_newProxy(t *testing.T) {
 				if !reflect.DeepEqual(found, testBytes) {
 					t.Errorf("expected body '%v', got '%v'", testBytes, found)
 				}
+			}
+		})
+	}
+}
+
+func Test_stripLeadingPaths(t *testing.T) {
+	type args struct {
+		path string
+	}
+	tests := []struct {
+		name string
+		args args
+		want string
+	}{
+		{"remove prefix", args{"/networks/test_network/api/version"}, "/version"},
+		{"no prefix removed", args{"/api/version"}, "/api/version"},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := stripLeadingSegments(tt.args.path); got != tt.want {
+				t.Errorf("stripLeadingPaths() = %v, want %v", got, tt.want)
 			}
 		})
 	}
